refactor: extract supportsSyncHooks into an import-free module
`supportsSyncHooks` is a pure Node.js version check, but it lived in
create-hook.mjs alongside the loader's acorn / cjs-module-lexer dependency
graph. A CommonJS preloader that wants to decide whether to load the ESM
synchronous loader at all had to import all of that just to call the check,
loading the transformer machinery on the main thread on runtimes where the
synchronous loader cannot even run.

Move the function to a new import-free supports-sync-hooks.mjs and re-export it
from create-hook.mjs and register-hooks.mjs, so every existing import path keeps
working while consumers can also import the predicate on its own.
